Create the perma id for IDO
The Imperial Data Ontology (IDO) is an OWL 2 DL ontology that allows the description of selling data of fashion design industries.

The development of IDO is one of the outcomes of the agreement between the [University of Bologna](http://www.unibo.it/en) and [Imperial](http://www.imperialfashion.com/en/). Part of this agreement is dedicated to the use of Semantic Web technologies for the description and analysis of selling data of Imperial, and, more generally, of fashion design industries.