Fixed .htaccess for versions and sources
The .htaccess has been modified so as to provide a correct mechanism for the content negotiation. 

When accessed by means of a browser (i.e. requesting HTML as format), the current version of UNDO redirects to the official UN document describing the ontology, available at https://unsceb-hlcm.github.io/onto-undo/index.html. Any other request (i.e. different formats and version IRI of the ontology) will be redirected to the specific file contained in the repository.