Create new permanent id for RASH and its Framework
This permanent ID space has been created for providing a way to permanently identify all the resources associated to the RASH Framework, i.e. a set of specifications and writing/conversion/extraction tools for writing academic articles in RASH, i.e., a markup language defined as a subset of HTML for writing scientific articles.

The responsible for this space is Silvio Peroni, who can be contacted by email (essepuntato@gmail.com) or twitter (@essepuntato).