Updated redirection rules.
For consistency with `https://w3id.org/emmo` that cant be distinguished
from `https://w3id.org/emmo/`, URLs with and without trailing slash are
now equivalent.

This is probably also what people would expect, but makes it slightly
less convinient to refer to the source files in the github repo.