Fix: trailing slash and lookup order for FEP documents
URLs of the form https://w3id.org/fep/888d/ used to land you on the FEP folder instead of the FEP document. This issue was made worse by some autolinkers inserting a / before a fragment identifier. Fix this so that FEP documents are looked up before subresources in the case of a single trailing slash.