Redirect to catena-X odrl.jsonld
As we limited the allowed scope of odrl in this file https://github.com/catenax-eV/cx-odrl-profile/blob/main/schema/context/odrl.jsonld and thus need to reference this one in the constraints in CX-0152. 

New rule and sequence was tested in https://htaccess.madewithlove.com/ . Please remove in the "url that you are applying the rules to" /catenax.